@@ -2,9 +2,7 @@ package bui.quocdat.dchat.Activity;
 
 import android.annotation.SuppressLint;
 import android.content.ContentResolver;
-import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.net.Uri;
 import android.os.Bundle;
 import android.text.TextUtils;
@@ -17,7 +15,6 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
@@ -26,11 +23,9 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
-import com.github.nkzawa.emitter.Emitter;
 import com.github.nkzawa.socketio.client.Socket;
 import com.google.android.gms.tasks.Continuation;
 import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.Task;
 import com.google.android.material.bottomsheet.BottomSheetDialog;
 import com.google.android.material.card.MaterialCardView;
@@ -46,9 +41,11 @@ import org.json.JSONObject;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 
 import bui.quocdat.dchat.Adapter.MessageAdapter;
 import bui.quocdat.dchat.Other.Message;
+import bui.quocdat.dchat.Other.PreferenceManager;
 import bui.quocdat.dchat.Other.Strings;
 import bui.quocdat.dchat.R;
 import bui.quocdat.dchat.Socketconnetion.SocketManager;
@@ -106,8 +103,7 @@ public class ChatActivity extends AppCompatActivity implements View.OnClickListe
         Intent intent = getIntent();
         userIDFriend = String.valueOf(intent.getIntExtra("userID", 1));
 
-        SharedPreferences sharedPreferences = getSharedPreferences(Strings.SHARED_PREFERENCES_NAME, Context.MODE_PRIVATE);
-        id = sharedPreferences.getString(Strings.USER_ID, "");
+        id = new PreferenceManager(this).getString(Strings.USER_ID);
 
         //initView
         initView();
@@ -115,7 +111,7 @@ public class ChatActivity extends AppCompatActivity implements View.OnClickListe
         //Toolbar
         Toolbar toolbar = findViewById(R.id.chat_toolbar);
         setSupportActionBar(toolbar);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        Objects.requireNonNull(getSupportActionBar()).setDisplayHomeAsUpEnabled(true);
 
         //recyclerview
         recyclerView = findViewById(R.id.recycler_view_message);
@@ -125,32 +121,24 @@ public class ChatActivity extends AppCompatActivity implements View.OnClickListe
         recyclerView.setLayoutManager(layoutManager);
 
 
-        socket.emit("getInfOfUser", userIDFriend).on("resultInfOfUser", new Emitter.Listener() {
-            @Override
-            public void call(final Object... args) {
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        JSONObject infUser = (JSONObject) args[0];
-                        try {
-                            String url = infUser.getString("url");
-                            if (!url.isEmpty()) {
-                                Glide.with(getApplicationContext())
-                                        .load(url)
-                                        .into(imgUser);
-                            }
-                            useName.setText(infUser.getString("last_name"));
-                            if (infUser.getBoolean("status")) tvStatus.setText("Online");
-                            else tvStatus.setText("Offline");
-                            imageUrl = infUser.getString("url");
-                            readMessage(id, userIDFriend, infUser.getString("url"));
-                        } catch (JSONException e) {
-                            e.printStackTrace();
-                        }
-                    }
-                });
+        socket.emit("getInfOfUser", userIDFriend).on("resultInfOfUser", args -> runOnUiThread(() -> {
+            JSONObject infUser = (JSONObject) args[0];
+            try {
+                String url = infUser.getString("url");
+                if (!url.isEmpty()) {
+                    Glide.with(getApplicationContext())
+                            .load(url)
+                            .into(imgUser);
+                }
+                useName.setText(String.format("%s %s", infUser.getString("first_name"), infUser.getString("last_name")));
+                if (infUser.getBoolean("status")) tvStatus.setText("Online");
+                else tvStatus.setText("Offline");
+                imageUrl = infUser.getString("url");
+                readMessage(id, userIDFriend, infUser.getString("url"));
+            } catch (JSONException e) {
+                e.printStackTrace();
             }
-        });
+        }));
 
 
         // set Even click
@@ -163,31 +151,18 @@ public class ChatActivity extends AppCompatActivity implements View.OnClickListe
         imgUser.setOnClickListener(this);
         findViewById(R.id.img_attach).setOnClickListener(this);
 
-        add_close.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                uriMess = null;
-                add_layout.setVisibility(View.GONE);
-                isPicture = false;
-                isFile = false;
-                isLocation = false;
-                isMusic = false;
-            }
+        add_close.setOnClickListener(view -> {
+            uriMess = null;
+            add_layout.setVisibility(View.GONE);
+            isPicture = false;
+            isFile = false;
+            isLocation = false;
+            isMusic = false;
         });
 
         markAsRead();
 
-        socket.on("result_mark", new Emitter.Listener() {
-            @Override
-            public void call(Object... args) {
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        readMessage(id, userIDFriend, imageUrl);
-                    }
-                });
-            }
-        });
+        socket.on("result_mark", args -> runOnUiThread(() -> readMessage(id, userIDFriend, imageUrl)));
 
     }
 
@@ -214,40 +189,32 @@ public class ChatActivity extends AppCompatActivity implements View.OnClickListe
     }
 
     private void readMessage(String myId, String userID, final String imageUrl) {
-        socket.emit("getMessages", myId, userID).on("resultGetMessages", new Emitter.Listener() {
-            @Override
-            public void call(final Object... args) {
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        JSONArray array = (JSONArray) args[0];
-                        listData.clear();
-                        try {
-                            Message message;
-                            JSONObject current;
-                            for (int i = 0; i < array.length(); i++) {
-                                current = array.getJSONObject(i);
-                                message = new Message(Integer.parseInt(current.getString("id"))
-                                        , Integer.parseInt(current.getString("sender_id"))
-                                        , Integer.parseInt(current.getString("receiver_id"))
-                                        , current.getString("text")
-                                        , current.getString("created_at")
-                                        , current.getString("type")
-                                        , current.getString("url")
-                                        , current.getBoolean("is_seen"));
-                                listData.add(message);
-                            }
-                            Collections.reverse(listData);
-                            messageAdapter = new MessageAdapter(listData, imageUrl, getApplicationContext());
-                            messageAdapter.notifyDataSetChanged();
-                            recyclerView.setAdapter(messageAdapter);
-                        } catch (JSONException e) {
-                            e.printStackTrace();
-                        }
-                    }
-                });
+        socket.emit("getMessages", myId, userID).on("resultGetMessages", args -> runOnUiThread(() -> {
+            JSONArray array = (JSONArray) args[0];
+            listData.clear();
+            try {
+                Message message;
+                JSONObject current;
+                for (int i = 0; i < array.length(); i++) {
+                    current = array.getJSONObject(i);
+                    message = new Message(Integer.parseInt(current.getString("id"))
+                            , Integer.parseInt(current.getString("sender_id"))
+                            , Integer.parseInt(current.getString("receiver_id"))
+                            , current.getString("text")
+                            , current.getString("created_at")
+                            , current.getString("type")
+                            , current.getString("url")
+                            , current.getBoolean("is_seen"));
+                    listData.add(message);
+                }
+                Collections.reverse(listData);
+                messageAdapter = new MessageAdapter(listData, imageUrl, getApplicationContext());
+                messageAdapter.notifyDataSetChanged();
+                recyclerView.setAdapter(messageAdapter);
+            } catch (JSONException e) {
+                e.printStackTrace();
             }
-        });
+        }));
     }
 
     private void setStatus(Boolean isOnline) {
@@ -261,39 +228,31 @@ public class ChatActivity extends AppCompatActivity implements View.OnClickListe
         super.onResume();
         setStatus(true);
         //received the message
-        socket.on("messagefromsingle", new Emitter.Listener() {
-            @Override
-            public void call(final Object... args) {
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        JSONObject data = (JSONObject) args[0];
-                        Message message;
-                        try {
-                            message = new Message(Integer.parseInt(data.getString("id"))
-                                    , Integer.parseInt(data.getString("sender_id"))
-                                    , Integer.parseInt(data.getString("receiver_id"))
-                                    , data.getString("text")
-                                    , data.getString("created_at")
-                                    , data.getString("type")
-                                    , data.getString("url")
-                                    , data.getBoolean("is_seen"));
-                            listData.add(message);
-                            messageAdapter = new MessageAdapter(listData, imageUrl, getApplicationContext());
-                            messageAdapter.notifyDataSetChanged();
-                            recyclerView.setAdapter(messageAdapter);
-                            Log.d("messages", "go");
-                            if (!id.equals(String.valueOf(data.getInt("sender_id")))){
-                                markAsRead();
-                                Log.d("messages", "mark");
-                            }
-                        } catch (JSONException e) {
-                            e.printStackTrace();
-                        }
-                    }
-                });
+        socket.on("messagefromsingle", args -> runOnUiThread(() -> {
+            JSONObject data = (JSONObject) args[0];
+            Message message;
+            try {
+                message = new Message(Integer.parseInt(data.getString("id"))
+                        , Integer.parseInt(data.getString("sender_id"))
+                        , Integer.parseInt(data.getString("receiver_id"))
+                        , data.getString("text")
+                        , data.getString("created_at")
+                        , data.getString("type")
+                        , data.getString("url")
+                        , data.getBoolean("is_seen"));
+                listData.add(message);
+                messageAdapter = new MessageAdapter(listData, imageUrl, getApplicationContext());
+                messageAdapter.notifyDataSetChanged();
+                recyclerView.setAdapter(messageAdapter);
+                Log.d("messages", "go");
+                if (!id.equals(String.valueOf(data.getInt("sender_id")))){
+                    markAsRead();
+                    Log.d("messages", "mark");
+                }
+            } catch (JSONException e) {
+                e.printStackTrace();
             }
-        });
+        }));
     }
 
     @Override
@@ -448,36 +407,25 @@ public class ChatActivity extends AppCompatActivity implements View.OnClickListe
                     + "." + getFileExtension(uriMess));
 
             StorageTask uploadTask = fileReference.putFile(uriMess);
-            uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
-                @Override
-                public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
-                    if (!task.isSuccessful()) {
-                        throw task.getException();
-                    }
-                    return fileReference.getDownloadUrl();
+            uploadTask.continueWithTask((Continuation<UploadTask.TaskSnapshot, Task<Uri>>) task -> {
+                if (!task.isSuccessful()) {
+                    throw task.getException();
                 }
-            }).addOnCompleteListener(new OnCompleteListener<Uri>() {
-                @Override
-                public void onComplete(@NonNull Task<Uri> task) {
-                    if (task.isSuccessful()) {
-                        Uri downloadUri = task.getResult();
-                        assert downloadUri != null;
-                        final String mUri = downloadUri.toString();
+                return fileReference.getDownloadUrl();
+            }).addOnCompleteListener((OnCompleteListener<Uri>) task -> {
+                if (task.isSuccessful()) {
+                    Uri downloadUri = task.getResult();
+                    assert downloadUri != null;
+                    final String mUri = downloadUri.toString();
 
-                        int code = (isFile) ? 402 : 400;
-                        socket.emit("messagedectectionfromsingle", code, object, mUri);
-                        uriMess = null;
-                    } else {
-                        Toast.makeText(ChatActivity.this, "Failed", Toast.LENGTH_SHORT).show();
-                    }
+                    int code = (isFile) ? 402 : 400;
+                    socket.emit("messagedectectionfromsingle", code, object, mUri);
+                    uriMess = null;
+                } else {
+                    Toast.makeText(ChatActivity.this, "Failed", Toast.LENGTH_SHORT).show();
+                }
 
-                }
-            }).addOnFailureListener(new OnFailureListener() {
-                @Override
-                public void onFailure(@NonNull Exception e) {
-                    Toast.makeText(ChatActivity.this, e.getMessage(), Toast.LENGTH_SHORT).show();
-                }
-            });
+            }).addOnFailureListener(e -> Toast.makeText(ChatActivity.this, e.getMessage(), Toast.LENGTH_SHORT).show());
         }
     }
 
