@@ -1,29 +1,19 @@
 package bui.quocdat.dchat.Activity;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.app.Activity;
-import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
 
+import androidx.appcompat.app.AppCompatActivity;
 
-import com.github.nkzawa.socketio.client.Socket;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-
-import java.util.Objects;
-
+import bui.quocdat.dchat.Other.PreferenceManager;
 import bui.quocdat.dchat.Other.Strings;
 import bui.quocdat.dchat.R;
 import bui.quocdat.dchat.Socketconnetion.SocketManager;
 
 public class StartActivity extends AppCompatActivity {
-
-    private SharedPreferences sharedPreferences;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -32,17 +22,12 @@ public class StartActivity extends AppCompatActivity {
 
         //socket
         SocketManager.getInstance().connectSocket();
-        init();
-        if (!Objects.requireNonNull(sharedPreferences.getString(Strings.STATUS, "false")).equals("false")){
+
+        PreferenceManager preferenceManager = new PreferenceManager(this);
+        if (preferenceManager.getBoolean(Strings.STATUS)){
             startActivity(new Intent(StartActivity.this, MainActivity.class));
             finish();
         }
-
-    }
-
-    private void init() {
-        sharedPreferences = getSharedPreferences(Strings.SHARED_PREFERENCES_NAME, Context.MODE_PRIVATE);
-
     }
 
     //hind keyboard
